requirements.txt
aerich==0.7.1
asyncpg==0.27.0
bcrypt==4.0.1
passlib==1.7.4
fastapi==0.90.0
python-jose==3.3.0
python-multipart==0.0.5
tortoise-orm==0.19.2
uvicorn==0.20.0
requests
minio
Fiona==1.8.21
geopandas==0.11.0
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
Shapely==1.8.2
pyproj==3.3.1
PyJWT==2.6.0
GDAL==3.2.2
rasterio==1.2.6
rioxarray==0.7.0
matplotlib==3.5.2
Rtree==0.9.7
xarray==0.19.0
netCDF4==1.5.7